Modifies navigation link for consistency
The previous version stated "Apply now", which is not consistent
with "The Program" and "Our Students". It was changed simply to
"Apply Now".

Recommended future work is to style "Apply Now" as a button so that
it is visually distinct from the other links, and to represent
a strong, call-to-action.